refactor(queue): extract job-dispatch.ts from processors.ts
Final step of #4013. Moves the top-level processJob dispatcher (a pure
switch over JobMessage["type"]) into its own file. This is the most
interdependent piece of the split -- job-dispatch.ts imports most of its
handlers back from processors.ts (which stays large; nothing else in
processors.ts functions (and processGitHubWebhook) are exported for this
one-directional import-back. processors.ts re-exports processJob for
src/index.ts, src/server.ts, and the test suite.

Also removes ~28 imports in processors.ts that were orphaned once
processJob (their only remaining caller) moved out, and adds two
regression tests for job-dispatch branches that were untested even in
the original file (retry-orb-relay dispatch; a notify-evaluate payload
with neither the events array nor the legacy singular event field).
